feat: made direction a enum class instead of a standard enum. changed room to use std::array

+++ core/Room.cpp
@@ -5,11 +5,11 @@ Room::Room(int roomNo)
     : _sides{nullptr, nullptr, nullptr, nullptr}, _roomNumber(roomNo) {}
 
 MapSite* Room::GetSide(Direction d) const {
-    return _sides[d];
+    return _sides[static_cast<size_t>(d)];
 }
 
 void Room::SetSide(Direction d, MapSite* site) {
-    _sides[d] = site;
+    _sides[static_cast<size_t>(d)] = site;
 }
 
 void Room::Enter() {
